requirements.txt




# FastAPI and core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0

# Vector database
qdrant-client>=1.9.0

# Azure OpenAI (for search queries only)
openai>=1.0.0
azure-core>=1.29.0

# LangChain for advanced summarization
langchain>=0.0.312
langchain-openai>=0.0.1
langchain-community>=0.0.10

# LLM Monitoring and Observability
langfuse>=2.6.1
langfuse-langchain>=0.2.4

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# Monitoring and Observability
applicationinsights>=0.11.10

# Optional: Data processing
numpy>=1.26.0
pytz>=2023.3

# Optional: Health monitoring
psutil>=5.9.6
langfuse
